apps/anthaathi-commerce-mobile-client/src/features/CMS/containers/Core/components/CMSText/index.spec.tsx
import React from 'react';
import 'react-native';
import {render} from '@testing-library/react-native';
import {IntlProvider} from 'react-intl';

import locale from '../../../../../../compiled-locales/en-US.json';
import {ThemeProvider} from 'react-native-paper';
import CMSText from '.';

describe('CMSText', () => {
  it('should render the item', function () {
    const temp = render(
      <ThemeProvider>
        <IntlProvider locale="en-US" messages={locale}>
          <CMSText title="title" />
        </IntlProvider>
      </ThemeProvider>,
    );

    expect(temp).toMatchSnapshot();
    expect(temp.queryByTestId('CMSText')).toBeTruthy();
  });

  it('should have CMSText title', () => {
    const temp = render(
      <ThemeProvider>
        <IntlProvider locale="en-US" messages={locale}>
          <CMSText title="title" />
        </IntlProvider>
      </ThemeProvider>,
    );

    expect(temp.queryByText('title')).toBeTruthy();
  });
});
